+++ NEWS.md
@@ -1,7 +1,9 @@
 # qrlabelr 0.1.1
-* Added new options to print labels row-wise from left to right or column-wise from top to bottom.
-* Added new options to draw rectangles around labels or no rectangles.
-
+* Added new options to print labels row-wise from left to right or column-wise from top to bottom to customizable functions and Shiny app.
+* Added new options to draw rectangles around labels to customizable functions and Shiny app.
+* Simplified the Shiny app by disabling subset functionality.
+* Field plot label option in Shiny app can now show unique ID text on the label.
+* Fixed bug in Treetag label option in Shiny app.
 
 # qrlabelr 0.1.0